Introduce math.gl, break out mercator utils
