+++ .eslintrc
@@ -7,7 +7,8 @@
     "guard-for-in": 0,
     "no-inline-comments": 0,
     "max-statements": 0,
-    "no-multi-spaces": 0
+    "no-multi-spaces": 0,
+    "camelcase": 0
   }
 }
