+++ src/index.js
@@ -2,3 +2,11 @@
 export {default as default} from './flat-mercator-viewport';
 export {default as FlatMercatorViewport} from './flat-mercator-viewport';
 export {default as PerspectiveMercatorViewport} from './perspective-mercator-viewport';
+
+export {projectFlat} from './web-mercator-utils';
+export {unprojectFlat} from './web-mercator-utils';
+export {calculateDistanceScales} from './web-mercator-utils';
+export {getFov} from './web-mercator-utils';
+export {getClippingPlanes} from './web-mercator-utils';
+export {makeProjectionMatrixFromMercatorParams} from './web-mercator-utils';
+export {makeUncenteredViewMatrixFromMercatorParams} from './web-mercator-utils';
